Get series or series-group details

get_details returns metadata from a Bank of Canada series or series group.

get_details(name = NULL, group = FALSE)

Arguments

name

A character of length 1 indicating the series or series group for which information should be retrieved.

group

A Boolean indicating whether the name is a series or a series group.

Value

A list of series or group details.

Examples

get_details("CES_C4E_LOSE_JOB_SK")
#> $name
#> [1] "CES_C4E_LOSE_JOB_SK"
#> 
#> $label
#> [1] "Probability of losing job in the next 12 months (left scale), Saskatchewan"
#> 
#> $description
#> [1] "Probability of losing job in the next 12 months (left scale), Saskatchewan"
#> 
if (FALSE) { # \dontrun{
get_details("BAPF_TRANSACTION_DATA")
} # }
get_details("BAPF_TRANSACTION_DATA", group = TRUE)
#> $name
#> [1] "BAPF_TRANSACTION_DATA"
#> 
#> $label
#> [1] "Bankers’ Acceptance Purchase Facility, transaction level data"
#> 
#> $description
#> [1] "Bankers’ Acceptance Purchase Facility, transaction level data"
#> 
#> $groupSeries
#> $groupSeries$BAPF_TRANSACTION_TRADE_DATE
#> $groupSeries$BAPF_TRANSACTION_TRADE_DATE$label
#> [1] "Trade date"
#> 
#> $groupSeries$BAPF_TRANSACTION_TRADE_DATE$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_TRADE_DATE"
#> 
#> 
#> $groupSeries$BAPF_TRANSACTION_SETTLEMENT_DATE
#> $groupSeries$BAPF_TRANSACTION_SETTLEMENT_DATE$label
#> [1] "Settlement date"
#> 
#> $groupSeries$BAPF_TRANSACTION_SETTLEMENT_DATE$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_SETTLEMENT_DATE"
#> 
#> 
#> $groupSeries$BAPF_TRANSACTION_AMOUNT_PURCHASED
#> $groupSeries$BAPF_TRANSACTION_AMOUNT_PURCHASED$label
#> [1] "Amount purchased($)"
#> 
#> $groupSeries$BAPF_TRANSACTION_AMOUNT_PURCHASED$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_AMOUNT_PURCHASED"
#> 
#> 
#> $groupSeries$BAPF_TRANSACTION_ISIN
#> $groupSeries$BAPF_TRANSACTION_ISIN$label
#> [1] "ISIN"
#> 
#> $groupSeries$BAPF_TRANSACTION_ISIN$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_ISIN"
#> 
#> 
#> $groupSeries$BAPF_TRANSACTION_MATURITY_DATE
#> $groupSeries$BAPF_TRANSACTION_MATURITY_DATE$label
#> [1] "Security maturity date"
#> 
#> $groupSeries$BAPF_TRANSACTION_MATURITY_DATE$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_MATURITY_DATE"
#> 
#> 
#> $groupSeries$BAPF_TRANSACTION_ISSUER_NAME
#> $groupSeries$BAPF_TRANSACTION_ISSUER_NAME$label
#> [1] "Issuer name"
#> 
#> $groupSeries$BAPF_TRANSACTION_ISSUER_NAME$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_ISSUER_NAME"
#> 
#> 
#> $groupSeries$BAPF_TRANSACTION_PURCHASE_PRICE
#> $groupSeries$BAPF_TRANSACTION_PURCHASE_PRICE$label
#> [1] "Purchase price($)"
#> 
#> $groupSeries$BAPF_TRANSACTION_PURCHASE_PRICE$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_PURCHASE_PRICE"
#> 
#> 
#> $groupSeries$BAPF_TRANSACTION_PURCHASE_YIELD
#> $groupSeries$BAPF_TRANSACTION_PURCHASE_YIELD$label
#> [1] "Purchase yield(%)"
#> 
#> $groupSeries$BAPF_TRANSACTION_PURCHASE_YIELD$link
#> [1] "https://www.bankofcanada.ca/valet/series/BAPF_TRANSACTION_PURCHASE_YIELD"
#> 
#> 
#>