Get series-group observations

get_group returns observations from a Bank of Canada series group.

get_group(name = NULL, ...)

Arguments

name

A character of length 1 indicating the series group to retrieve.

...

Additional query parameters. Possible values are start_date and/or end_date (both character), or one of recent, recent_weeks, recent_months, or recent_years (all numeric).

Value

A tibble.

Details

Valet, the server-side API, does not always return observations filtered by ... arguments for series groups, even if it will accept the request.

Examples

get_group("BAPF_TRANSACTION_DATA")
#> # A tibble: 201 × 9
#>    id    trade_date settlement_date amount_purchased isin         maturity_date
#>    <chr> <date>     <date>                     <dbl> <chr>        <date>       
#>  1 1     2020-03-23 2020-03-25             250000000 CA3379Z2RA26 2020-04-01   
#>  2 2     2020-03-23 2020-03-25             200000000 CA3379Z2R389 2020-04-03   
#>  3 3     2020-03-23 2020-03-25             500000000 CA3379Z2RF13 2020-04-14   
#>  4 4     2020-03-23 2020-03-25             500000000 CA8911Z1RM36 2020-04-20   
#>  5 5     2020-03-23 2020-03-25             300000000 CA3379Z2RE48 2020-04-13   
#>  6 6     2020-03-23 2020-03-25             280000000 CA3379Z2RH78 2020-04-16   
#>  7 7     2020-03-23 2020-03-25             460000000 CA3379Z2RJ35 2020-04-17   
#>  8 8     2020-03-23 2020-03-25             680000000 CA3379Z2RM63 2020-04-20   
#>  9 9     2020-03-23 2020-03-25             350000000 CA3379Z2RP94 2020-04-22   
#> 10 10    2020-03-23 2020-03-25             125000000 CA3379Z2RQ77 2020-04-23   
#> # … with 191 more rows, and 3 more variables: issuer_name <chr>,
#> #   purchase_price <dbl>, purchase_yield <dbl>
get_group("gbpp")
#> # A tibble: 3,600 × 17
#>    id        auction_date total_amount settlement_date maturity_date coupon_rate
#>    <chr>     <date>              <dbl> <date>          <date>              <dbl>
#>  1 2022-04-… 2022-04-07      300000000 2022-04-11      NA                  NA   
#>  2 2022-04-… 2022-04-07             NA 2022-04-11      2028-06-01           2   
#>  3 2022-04-… 2022-04-07             NA 2022-04-11      2029-06-01           2.25
#>  4 2022-04-… 2022-04-07             NA 2022-04-11      2029-06-01           5.75
#>  5 2022-04-… 2022-04-07             NA 2022-04-11      2030-06-01           1.25
#>  6 2022-04-… 2022-04-07             NA 2022-04-11      2030-12-01           0.5 
#>  7 2022-04-… 2022-04-07             NA 2022-04-11      2031-06-01           1.5 
#>  8 2022-04-… 2022-04-07             NA 2022-04-11      2031-12-01           1.5 
#>  9 2022-04-… 2022-04-07             NA 2022-04-11      2033-06-01           5.75
#> 10 2022-04-… 2022-04-07             NA 2022-04-11      2037-06-01           5   
#> # … with 3,590 more rows, and 11 more variables: isin <chr>,
#> #   interest_start_date <date>, interest_end_date <date>, interest_rate <dbl>,
#> #   auction_type <chr>, total_amount_purchased <dbl>, amount_purchased <dbl>,
#> #   cutoff_yield <dbl>, allotment_ratio <dbl>, deadline <time>, term <chr>

if (FALSE) {
#this is a series
get_group("FXCADAUD")
}